+++ app/src/main/java/com/github/se/gomeet/ui/mainscreens/events/manageinvites/ManageInvites.kt
@@ -1,31 +1,22 @@
-package com.github.se.gomeet.ui.mainscreens.events
+package com.github.se.gomeet.ui.mainscreens.events.manageinvites
 
-import android.util.Log
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.PagerState
 import androidx.compose.foundation.pager.rememberPagerState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.KeyboardArrowLeft
-import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonColors
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.CircularProgressIndicator
@@ -41,29 +32,22 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import coil.compose.rememberAsyncImagePainter
-import coil.request.ImageRequest
-import com.github.se.gomeet.R
 import com.github.se.gomeet.model.event.Event
 import com.github.se.gomeet.model.event.Invitation
 import com.github.se.gomeet.model.event.InviteStatus
-import com.github.se.gomeet.model.event.InviteStatus.*
+import com.github.se.gomeet.model.event.InviteStatus.ACCEPTED
+import com.github.se.gomeet.model.event.InviteStatus.PENDING
+import com.github.se.gomeet.model.event.InviteStatus.REFUSED
+import com.github.se.gomeet.model.event.InviteStatus.TO_INVITE
 import com.github.se.gomeet.model.user.GoMeetUser
 import com.github.se.gomeet.ui.navigation.BottomNavigationMenu
 import com.github.se.gomeet.ui.navigation.NavigationActions
@@ -171,7 +155,8 @@ fun ManageInvites(
                           text = TO_INVITE.formattedName,
                           style =
                               MaterialTheme.typography.bodyMedium.copy(
-                                  fontWeight = pagerWeight(pagerState, TO_INVITE.formattedName)))
+                                  fontWeight = pagerWeight(pagerState, TO_INVITE.formattedName)
+                              ))
                     }
 
                 // Pending invitations
@@ -185,7 +170,8 @@ fun ManageInvites(
                           text = PENDING.formattedName,
                           style =
                               MaterialTheme.typography.bodyMedium.copy(
-                                  fontWeight = pagerWeight(pagerState, PENDING.formattedName)))
+                                  fontWeight = pagerWeight(pagerState, PENDING.formattedName)
+                              ))
                     }
 
                 // Accepted invitations
@@ -199,7 +185,8 @@ fun ManageInvites(
                           text = ACCEPTED.formattedName,
                           style =
                               MaterialTheme.typography.bodyMedium.copy(
-                                  fontWeight = pagerWeight(pagerState, ACCEPTED.formattedName)))
+                                  fontWeight = pagerWeight(pagerState, ACCEPTED.formattedName)
+                              ))
                     }
 
                 // Refused invitations
@@ -213,7 +200,8 @@ fun ManageInvites(
                           text = REFUSED.formattedName,
                           style =
                               MaterialTheme.typography.bodyMedium.copy(
-                                  fontWeight = pagerWeight(pagerState, REFUSED.formattedName)))
+                                  fontWeight = pagerWeight(pagerState, REFUSED.formattedName)
+                              ))
                     }
               }
           Canvas(
@@ -304,139 +292,6 @@ fun ManageInvites(
       }
 }
 
-/**
- * This composable function represents the list of users that can be invited to an event.
- *
- * @param list the list of users that can be invited
- * @param currentEvent the event for which the users are invited
- * @param status the status of the invitation
- * @param callback the callback function to update the list of users to invite
- * @param initialClicked the initial state of the button
- */
-@Composable
-fun PageUserInvites(
-    list: List<GoMeetUser>,
-    currentEvent: Event,
-    status: InviteStatus?,
-    callback: (GoMeetUser) -> Unit,
-    initialClicked: Boolean
-) {
-
-  Column(
-      verticalArrangement = Arrangement.Top,
-      horizontalAlignment = Alignment.CenterHorizontally,
-      modifier = Modifier.fillMaxSize()) {
-        list.forEach { follower ->
-          UserInviteWidget(follower, currentEvent, status, initialClicked, callback)
-        }
-      }
-}
-
-/**
- * This composable function represents the user invite widget when the user can be invited to an
- * event.
- *
- * @param user the user that can be invited
- * @param event the event for which the user is invited
- * @param status the status of the invitation
- * @param initialClicked the initial state of the button
- * @param callback the callback function to update the list of users to invite
- */
-@Composable
-fun UserInviteWidget(
-    user: GoMeetUser,
-    event: Event,
-    status: InviteStatus?,
-    initialClicked: Boolean,
-    callback: (GoMeetUser) -> Unit
-) {
-
-  var clicked by rememberSaveable { mutableStateOf(initialClicked) }
-  Row(
-      modifier =
-          Modifier.fillMaxWidth()
-              .padding(start = 15.dp, end = 15.dp)
-              .height(50.dp)
-              .testTag("UserInviteWidget"),
-      horizontalArrangement = Arrangement.SpaceBetween,
-      verticalAlignment = Alignment.CenterVertically) {
-        val painter: Painter =
-            if (event.images.isNotEmpty()) {
-              rememberAsyncImagePainter(
-                  ImageRequest.Builder(LocalContext.current)
-                      .data(data = user.profilePicture)
-                      .apply {
-                        crossfade(true)
-                        placeholder(R.drawable.gomeet_logo)
-                      }
-                      .build())
-            } else {
-              painterResource(id = R.drawable.gomeet_logo)
-            }
-        // Profile picture
-        Image(
-            modifier =
-                Modifier.size(40.dp)
-                    .clip(CircleShape)
-                    .background(color = MaterialTheme.colorScheme.background),
-            painter = painter,
-            contentDescription = "profile picture",
-            contentScale = ContentScale.None)
-
-        // Username text
-        Text(text = user.username, color = MaterialTheme.colorScheme.onBackground)
-
-        // Status text
-        Text(
-            text =
-                when (status) {
-                  PENDING -> PENDING.formattedName
-                  ACCEPTED -> ACCEPTED.formattedName
-                  REFUSED -> REFUSED.formattedName
-                  else -> ""
-                },
-            modifier = Modifier.width(80.dp).testTag("InviteStatus"),
-            color =
-                when (status) {
-                  PENDING ->
-                      if (clicked) MaterialTheme.colorScheme.onBackground
-                      else MaterialTheme.colorScheme.primary
-                  ACCEPTED -> Color.Green
-                  REFUSED -> Color.Red
-                  else -> MaterialTheme.colorScheme.onBackground
-                })
-
-        // Button to invite or cancel invitation
-        Button(
-            onClick = {
-              clicked = !clicked
-              val toAdd =
-                  (clicked && (status == null || status == REFUSED)) ||
-                      (!clicked && (status == PENDING || status == ACCEPTED))
-              Log.d("ManageInvites", "toAdd: $toAdd, clicked: $clicked, status: $status")
-              callback(user.copy(pendingRequests = pendingRequests(user, event, status, toAdd)))
-            },
-            modifier = Modifier.height(26.dp).width(82.dp),
-            contentPadding = PaddingValues(vertical = 2.dp),
-            shape = RoundedCornerShape(10.dp),
-            colors = manageInvitesButtonColour(status, clicked)) {
-              Text(
-                  text =
-                      when (status) {
-                        PENDING,
-                        ACCEPTED -> if (clicked) "Invite" else "Cancel"
-                        else -> if (!clicked) "Invite" else "Cancel"
-                      },
-                  color =
-                      when (status) {
-                        PENDING,
-                        ACCEPTED -> if (clicked) Color.White else Color.DarkGray
-                        else -> if (!clicked) Color.White else Color.DarkGray
-                      },
-                  fontSize = 12.sp)
-            }
-      }
-}
 
 /**
  * This function updates the list of users to invite to an event.
@@ -527,7 +382,7 @@ fun pendingRequests(
  * @return The button colours.
  */
 @Composable
-private fun manageInvitesButtonColour(status: InviteStatus?, clicked: Boolean): ButtonColors {
+internal fun manageInvitesButtonColour(status: InviteStatus?, clicked: Boolean): ButtonColors {
 
   val c1 = MaterialTheme.colorScheme.outlineVariant
   val c2 = Color.LightGray
